Title: 최댓값 구하기, Time: 140ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 241108/최댓값 구하기/maximum-value.java
@@ -8,7 +8,7 @@ public class Main {
         int b = sc.nextInt();
         int c = sc.nextInt();
 
-        int avg = (a+b+c)/3;
+        double avg = (a+b+c)/3;
 
         if (avg > a){
             if (b>=c)
